# deploy/openapi.yaml
openapi: 3.0.3
info:
  title: 重明 (ChongMing) API
  description: |
    重明 AI 自动化测试平台 API 文档
    
    ## 核心模块
    - **神经设计层**: PRD 解析、用例生成
    - **左瞳引擎**: API 测试执行
    - **右瞳引擎**: UI 测试执行
    - **涡轮引擎**: 性能测试
    - **凤凰涅槃层**: 脚本编译
    - **缺陷分析**: 根因分析、修复建议
    
  version: 2.0.0
  contact:
    name: ChongMing Team
    email: chongming@example.com
  license:
    name: GNU AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8000/api/v1
    description: 开发环境
  - url: https://test-api.chongming.ai/api/v1
    description: 测试环境
  - url: https://api.chongming.ai/api/v1
    description: 生产环境
...
paths:
...
  /test-cases:
    get:
      tags: [test-cases]
      summary: 获取用例列表
      operationId: listTestCases
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: mode
          in: query
          schema:
            type: string
            enum: [UI, API, HYBRID]
        - name: priority
          in: query
          schema:
            type: string
            enum: [P0, P1, P2]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestCaseList'

    post:
      tags: [test-cases]
      summary: 创建测试用例
      operationId: createTestCase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TCIR'
      responses:
        '201':
          description: 创建成功
  /test-cases/{tc_id}:
    get:
      tags: [test-cases]
      summary: 获取用例详情
      operationId: getTestCase
      parameters:
        - name: tc_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TCIR'

  # ==================== 测试执行 ====================
...
components:
  schemas:
    TCIR:
      type: object
      description: 测试用例中间表示
      properties:
        protocol:
          type: string
          default: TC-IR
        version:
          type: string
          default: "1.0"
        id:
          type: string
        name:
          type: string
        mode:
          type: string
          enum: [UI, API, HYBRID]
        priority:
          type: string
          enum: [P0, P1, P2]
        steps:
          type: array
          items:
            type: object
            properties:
              action:
                type: string
              target:
                type: string
              value:
                type: string
        assertions:
          type: array
          items:
            type: object
...
    TestCaseList:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/TCIR'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []